refactor: markdownit composable
