+++ src/composables/index.ts
@@ -1,9 +1,11 @@
 import useDebounce from './useDebounce'
 import useMarkdownActions from './useMarkdownActions'
+import useMarkdownIt from './useMarkdownIt'
 import useShikiji from './useShikiji'
 
 export default {
   useDebounce,
   useMarkdownActions,
+  useMarkdownIt,
   useShikiji,
 }
